Добавлена база данных в проект

+++ src/main/java/org/userservice/User.java
@@ -6,6 +6,6 @@ public class User {
     private int id;
     private String name;
     private String email;
-    private byte age;
+    private short age;
     private LocalDateTime created = LocalDateTime.now();
 }
